+++ hal/targets/hal/TARGET_STM/TARGET_STM32F3/TARGET_NUCLEO_F302R8/objects.h
@@ -96,21 +96,12 @@ struct i2c_s {
     uint32_t slave;
 };
 
-struct pwmout_s {
-    PWMName pwm;
-    PinName pin;
-    uint32_t prescaler;
-    uint32_t period;
-    uint32_t pulse;
-    uint32_t channel;
-    uint32_t inverted;
-};
-
 struct can_s {
     CANName can;
     int index;
 };
 
+#include "common_objects.h"
 #include "gpio_object.h"
 
 #ifdef __cplusplus
